[BUGFIX] Prevent duplicate of theodia places with free storage
It may happen that a theodia place has been moved "outside" of the
current site.

In that particular context we have:

- Pastoral Unit
  - default storage
  - Parish 1
    - storage

We are in the context of the Pastoral Unit, allowing free storage,
the place has been moved/created within Parish and we expect to
use it without having to get a duplicate in the default storage.